# schema/openapi/openapi.yaml
openapi: '3.0.3'
info:
  title: Backbone API
  version: '1.0.0'
  description: |
    Auto-generated by backbone-cli.
    This API provides 11 standard CRUD endpoints for each model.
servers:
  - url: http://localhost:3000
    description: Development server
...
paths:
...
  /api/v1/digest_tips:
    get:
      tags:
        - DigestTips
      summary: List DigestTips
      description: Get a paginated list of DigestTips with optional filtering and sorting
      operationId: listDigestTips
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/PerPageParam'
        - $ref: '#/components/parameters/SortByParam'
        - $ref: '#/components/parameters/SortOrderParam'
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DigestTipList'
        '401':
          $ref: '#/components/responses/Unauthorized'

    post:
      tags:
        - DigestTips
      summary: Create DigestTip
      description: Create a new DigestTip
      operationId: createDigestTip
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateDigestTipRequest'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DigestTip'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
  /api/v1/digest_tip_users:
    get:
      tags:
        - DigestTipUsers
      summary: List DigestTipUsers
      description: Get a paginated list of DigestTipUsers with optional filtering and sorting
      operationId: listDigestTipUsers
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/PerPageParam'
        - $ref: '#/components/parameters/SortByParam'
        - $ref: '#/components/parameters/SortOrderParam'
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DigestTipUserList'
        '401':
          $ref: '#/components/responses/Unauthorized'

    post:
      tags:
        - DigestTipUsers
      summary: Create DigestTipUser
      description: Create a new DigestTipUser
      operationId: createDigestTipUser
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateDigestTipUserRequest'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DigestTipUser'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
components:
  schemas:
...
    DigestTip:
      type: object
      required:
        - id
        - sequence
        - name
        - tip_description
        - metadata
      properties:
        id:
          type: string
          format: uuid
        sequence:
          type: integer
          format: int32
        name:
          type: string
          maxLength: 255
        tip_description:
          type: string
        group_key:
          nullable: true
          type: string
          maxLength: 120
        metadata:
          type: object
...
    CreateDigestTipRequest:
      type: object
      required:
        - sequence
        - name
        - tip_description
        - metadata
      properties:
        sequence:
          type: integer
          format: int32
        name:
          type: string
          maxLength: 255
        tip_description:
          type: string
        group_key:
          nullable: true
          type: string
          maxLength: 120
        metadata:
          type: object
...
    DigestTipList:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/DigestTip'
        meta:
          $ref: '#/components/schemas/PaginationMeta'
    DigestTipUser:
      type: object
      required:
        - id
        - tip_id
        - user_id
        - metadata
      properties:
        id:
          type: string
          format: uuid
        tip_id:
          type: string
          format: uuid
        user_id:
          type: string
          format: uuid
        metadata:
          type: object
...
    CreateDigestTipUserRequest:
      type: object
      required:
        - tip_id
        - user_id
        - metadata
      properties:
        tip_id:
          type: string
          format: uuid
        user_id:
          type: string
          format: uuid
        metadata:
          type: object
...
    DigestTipUserList:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/DigestTipUser'
        meta:
          $ref: '#/components/schemas/PaginationMeta'
    PaginationMeta:
      type: object
      properties:
        total:
          type: integer
          example: 100
        page:
          type: integer
          example: 1
        per_page:
          type: integer
          example: 20
        total_pages:
          type: integer
          example: 5
    Error:
      type: object
      required:
        - code
        - message
      properties:
        code:
          type: string
          example: VALIDATION_ERROR
        message:
          type: string
          example: Invalid input data
        details:
          type: object
          additionalProperties: true
  parameters:
...
    PageParam:
      name: page
      in: query
      schema:
        type: integer
        minimum: 1
        default: 1
    PerPageParam:
      name: per_page
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
    SortByParam:
      name: sort_by
      in: query
      schema:
        type: string
    SortOrderParam:
      name: sort_order
      in: query
      schema:
        type: string
        enum: [asc, desc]
        default: asc
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
security:
  - bearerAuth: []